Initial commit -- .net WMI process watcher
This code is some sample code to watch for processes via the Windows Management Interface.
This code is written in C-sharp (C#), utilizing the Windows Management interfaces.

It is possible to rewrite this in C or C++, using the C/C++ interfaces to WMI, however
that exercise is left to the user.